+++ magik-lint/src/main/java/nl/ramsolutions/sw/magik/lint/MagikLint.java
@@ -95,13 +95,8 @@ public class MagikLint {
    */
   void showChecks(final Writer writer, final boolean showDisabled)
       throws ReflectiveOperationException, IOException {
-    final Path overrideConfigPath = this.properties.getPropertyPath(MagikLint.KEY_OVERRIDE_CONFIG);
-    final MagikToolsProperties checksProperties =
-        overrideConfigPath != null
-            ? new MagikToolsProperties(overrideConfigPath)
-            : MagikToolsProperties.DEFAULT_PROPERTIES;
     final MagikChecksConfiguration checksConfig =
-        new MagikChecksConfiguration(CheckList.getChecks(), checksProperties);
+        new MagikChecksConfiguration(CheckList.getChecks(), this.properties);
     final Iterable<MagikCheckHolder> holders = checksConfig.getAllChecks();
     for (final MagikCheckHolder holder : holders) {
       final MagikCheckMetadata metadata = holder.getMetadata();