+++ CHANGES.md
@@ -18,6 +18,7 @@
 - Add UnsafeEvaluateInvocation check to Sonar way profile.
 - Speed up looking for definitions based on path.
 - Fix UseValueCompare check to handle other notations for numbers + handle floats.
+- Fix magik-lint/magik-typed-lint not using located configuration file when showing checks.
 - Several fixes.
 
 0.10.1 (2024-08-14)